IC-1358 PR Review comments
