+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/dto/SentReferralDTO.kt
@@ -11,6 +11,7 @@ class SentReferralDTO(
   val referenceNumber: String,
   val assignedTo: AuthUserDTO?,
   val referral: DraftReferralDTO,
+  val actionPlanId: UUID?,
 ) {
   companion object {
     fun from(referral: Referral): SentReferralDTO {
@@ -21,6 +22,7 @@ class SentReferralDTO(
         referenceNumber = referral.referenceNumber!!,
         assignedTo = referral.assignedTo?.let { AuthUserDTO.from(it) },
         referral = DraftReferralDTO.from(referral),
+        actionPlanId = referral.actionPlan?.id
       )
     }
   }
