+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/dto/SentReferralDTOTest.kt
@@ -74,7 +74,53 @@ class SentReferralDTOTest(@Autowired private val json: JacksonTester<SentReferra
           "interpreterLanguage": "french",
           "serviceUser": {"crn": "X123456"},
           "serviceProvider": {"name": "Provider"}
-        }
+        },
+        "actionPlanId": null
+      }
+    """
+    )
+  }
+
+  @Test
+  fun `sent referral includes all draft fields and action plan id`() {
+    val id = UUID.fromString("3B9ED289-8412-41A9-8291-45E33E60276C")
+    val createdAt = OffsetDateTime.parse("2020-12-04T10:42:43+00:00")
+    val sentAt = OffsetDateTime.parse("2021-01-13T21:57:13+00:00")
+    val sentBy = AuthUser("id", "source", "username")
+
+    val referral = SampleData.sampleReferral(
+      "X123456",
+      "Provider",
+      id = id,
+      createdAt = createdAt,
+      actionPlan = SampleData.sampleActionPlan()
+    )
+
+    referral.referenceNumber = "something"
+    referral.needsInterpreter = true
+    referral.interpreterLanguage = "french"
+    referral.sentAt = sentAt
+    referral.sentBy = sentBy
+
+    val out = json.write(SentReferralDTO.from(referral))
+    Assertions.assertThat(out).isEqualToJson(
+      """
+      {
+        "sentAt": "2021-01-13T21:57:13Z",
+        "sentBy": {
+          "username": "username",
+          "authSource": "source"
+        },
+        "referenceNumber": "something",
+        "referral": {
+          "id": "3b9ed289-8412-41a9-8291-45e33e60276c",
+          "createdAt": "2020-12-04T10:42:43Z",
+          "needsInterpreter": true,
+          "interpreterLanguage": "french",
+          "serviceUser": {"crn": "X123456"},
+          "serviceProvider": {"name": "Provider"}
+        },
+        "actionPlanId": "${referral.actionPlan?.id}"
       }
     """
     )
