--- src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/dto/SentReferralWithActionPlanDTOTest.kt
@@ -1,129 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.dto
-
-import org.assertj.core.api.Assertions
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertDoesNotThrow
-import org.junit.jupiter.api.assertThrows
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.json.JsonTest
-import org.springframework.boot.test.json.JacksonTester
-import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.entity.AuthUser
-import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.entity.SampleData
-import java.time.OffsetDateTime
-import java.util.UUID
-
-@JsonTest
-class SentReferralWithActionPlanDTOTest(@Autowired private val json: JacksonTester<SentReferralWithActionPlanIdDTO>) {
-  @Test
-  fun `sent referral requires reference number, sent timestamp, sentBy`() {
-    val referral = SampleData.sampleReferral("X123456", "Provider")
-    val timestamp = OffsetDateTime.now()
-
-    assertThrows<RuntimeException> {
-      SentReferralWithActionPlanIdDTO.from(referral)
-    }
-
-    referral.sentAt = timestamp
-    assertThrows<RuntimeException> {
-      SentReferralWithActionPlanIdDTO.from(referral)
-    }
-
-    referral.referenceNumber = "something"
-    assertThrows<RuntimeException> {
-      SentReferralWithActionPlanIdDTO.from(referral)
-    }
-
-    referral.sentBy = AuthUser("id", "source", "username")
-    assertDoesNotThrow { SentReferralWithActionPlanIdDTO.from(referral) }
-  }
-
-  @Test
-  fun `sent referral includes all draft fields and action plan id`() {
-    val id = UUID.fromString("3B9ED289-8412-41A9-8291-45E33E60276C")
-    val createdAt = OffsetDateTime.parse("2020-12-04T10:42:43+00:00")
-    val sentAt = OffsetDateTime.parse("2021-01-13T21:57:13+00:00")
-    val sentBy = AuthUser("id", "source", "username")
-
-    val referral = SampleData.sampleReferral(
-      "X123456",
-      "Provider",
-      id = id,
-      createdAt = createdAt,
-      actionPlan = SampleData.sampleActionPlan()
-    )
-
-    referral.referenceNumber = "something"
-    referral.needsInterpreter = true
-    referral.interpreterLanguage = "french"
-    referral.sentAt = sentAt
-    referral.sentBy = sentBy
-
-    val out = json.write(SentReferralWithActionPlanIdDTO.from(referral))
-    Assertions.assertThat(out).isEqualToJson(
-      """
-      {
-        "sentAt": "2021-01-13T21:57:13Z",
-        "sentBy": {
-          "username": "username",
-          "authSource": "source"
-        },
-        "referenceNumber": "something",
-        "referral": {
-          "id": "3b9ed289-8412-41a9-8291-45e33e60276c",
-          "createdAt": "2020-12-04T10:42:43Z",
-          "needsInterpreter": true,
-          "interpreterLanguage": "french",
-          "serviceUser": {"crn": "X123456"},
-          "serviceProvider": {"name": "Provider"}
-        },
-        "actionPlanId": "${referral.actionPlan?.id}"
-      }
-    """
-    )
-  }
-
-  @Test
-  fun `sent referral includes all draft fields and action plan id is null when no plan exists`() {
-    val id = UUID.fromString("3B9ED289-8412-41A9-8291-45E33E60276C")
-    val createdAt = OffsetDateTime.parse("2020-12-04T10:42:43+00:00")
-    val sentAt = OffsetDateTime.parse("2021-01-13T21:57:13+00:00")
-    val sentBy = AuthUser("id", "source", "username")
-
-    val referral = SampleData.sampleReferral(
-      "X123456",
-      "Provider",
-      id = id,
-      createdAt = createdAt,
-      actionPlan = null
-    )
-
-    referral.referenceNumber = "something"
-    referral.needsInterpreter = true
-    referral.interpreterLanguage = "french"
-    referral.sentAt = sentAt
-    referral.sentBy = sentBy
-
-    val out = json.write(SentReferralWithActionPlanIdDTO.from(referral))
-    Assertions.assertThat(out).isEqualToJson(
-      """
-      {
-        "sentAt": "2021-01-13T21:57:13Z",
-        "sentBy": {
-          "username": "username",
-          "authSource": "source"
-        },
-        "referenceNumber": "something",
-        "referral": {
-          "id": "3b9ed289-8412-41a9-8291-45e33e60276c",
-          "createdAt": "2020-12-04T10:42:43Z",
-          "needsInterpreter": true,
-          "interpreterLanguage": "french",
-          "serviceUser": {"crn": "X123456"},
-          "serviceProvider": {"name": "Provider"}
-        },
-        "actionPlanId": null
-      }
-    """
-    )
-  }
-}
